small bits added before supporter skins, plus the official version for 1.20.1

+++ src/main/java/com/thetaciturnone/taccorpstrinkets/blocks/entities/TacPlushieBlockEntity.java
@@ -31,7 +31,7 @@ public class TacPlushieBlockEntity extends BlockEntity {
 	}
 
 	public void squish(int squash) {
-		this.squash += (double)squash;
+		this.squash += squash;
 		if (this.world != null) {
 			this.world.updateListeners(this.pos, this.getCachedState(), this.getCachedState(), 2);
 		}
